Forget about pre baking

+++ Assets/Scripts/Helper/PlaybackControlNonGUI.cs
@@ -4,8 +4,8 @@ using System.Collections;
 public class PlaybackControlNonGUI : MonoBehaviour {
 
 	public bool playing = true;
-	public decimal current_time;
-	public decimal total_time = 0;
+	public float current_time;
+	public float total_time = 0;
 
 	// Use this for initialization
 	void Start () {
@@ -18,7 +18,8 @@ public class PlaybackControlNonGUI : MonoBehaviour {
 		
 		if (playing) {
 			try {
-				current_time = (current_time + (decimal) Time.deltaTime) % total_time;
+				current_time = (current_time + Time.deltaTime) % total_time;
+				//Debug.Log(current_time);
 			} catch (System.DivideByZeroException) {
 				current_time = 0;
 			}
